+++ src/main/java/com/telepathicgrunt/the_bumblezone/entities/goals/BeehemothAIRide.java
@@ -4,6 +4,7 @@ import com.telepathicgrunt.the_bumblezone.configs.BzGeneralConfigs;
 import com.telepathicgrunt.the_bumblezone.entities.mobs.BeehemothEntity;
 import com.telepathicgrunt.the_bumblezone.mixin.entities.LivingEntityAccessor;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.ai.attributes.Attributes;
 import net.minecraft.entity.ai.goal.Goal;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.math.vector.Vector3d;
@@ -49,7 +50,7 @@ public class BeehemothAIRide extends Goal {
 
         if (player.zza != 0 || ((LivingEntityAccessor)player).isJumping()) {
             currentSpeed = Math.min(
-                    BzGeneralConfigs.beeheemothSpeed.get() * speedModifier,
+                    BzGeneralConfigs.beeheemothSpeed.get() * speedModifier * beehemothEntity.getFinalFlyingSpeed(),
                     currentSpeed + 0.3D);
         }
         else {